a facility for class slots to be unbound
for one of my projects, it has become useful to define a class which has slots
which are initially unbound and not passed in the constructor, but which can be
bound and accessed later. this commit adds a mechanism for defining such slots
to my `defclass` wrapper.

as part of this change, `defclass-boa-constructor` becomes intentionally less
flexible, no longer allowing its caller to supply a custom arglist or
constructor name. this likely doesn't matter, as that function is only and
should only ever be called by `defclass`.